requirements.txt
flask
flask-cors
openai>=1.0.0
pinecone-client>=3.0.0